Added interactive stuff
- Made a copy of the scene so we don't work on the same stuff all the
time.
- Added a script with the awesome name "Interactable" basically anything
we can click on and talk to should use this script or an extension of
it, items, npcs, stuff to use and look at. Anything with this script
should be tagged "Interactable" as well (could do this automatically)
- Changed to the PlayerMouseControl script to handle clicking on
anything interactable. If it's something we can use or pick up, it's set
as the player's target and he will move towards it and interact with it
when close enough.