Added the first steps towards click and drag
- The inventory panel shows when the mouse is in the top 1/8 part of the
screen
- Moved the logic in PlayerMouseControl around a little so we don't try
to walk around when we click the inventory
- Added some click and drag functionality, the InteractableRev class
handles actually moving the item around, the PlayerMouseControl handles
checking if we drop an item ontop of another.
STILL TODO: PC should walk towards target before checking whether a drop
does something. For now any item dropped on another will result in a
match and the item is removed from the inventory.