Fix for inventory collider
* On entering inventory, addItem() now changes the object's boxcollider
(only works with box collider, tests for one!) so the .center and .size
properties are, by default, set to 0,0,0, and 0.1,0.1,0.1 respectively.

This should create a box that roughly covers the area of the object. If
the object is more awkward, then the target collider variables are
public in the inspector.

* Changed settleItems() and settleItemsWithoutAnimation() to get target
y position from the attached InteractableRev.

* We should really change the name 'InteractableRev', I feel really
awkward about that now.